Fix for 2 config values
CONF_SCAN_INTERVAL and CONF_REPORT_SCAN_INTERVAL are saved in the data of entry, not in options.
This fix will check options first, check data second and fallback to the default if both are empty.

Fix for 2 config values

CONF_SCAN_INTERVAL and CONF_REPORT_SCAN_INTERVAL are saved in the data of entry, not in options.
This fix will check options first, check data second and fallback to the default if both are empty.